Change 'circled' to 'highlighted' in script
Looking at the script it seems that the word "highlighted" would fit better in this situation. The original word "circled" does not apply in a computers since you can't circle something in a computer. Hence why highlighted will be more fitting. Another saying like marked as important or selected could also be appropriate if you do not like highlighted.

Co-authored-by: Nathan <nchonot@nathannetwork.com>